Move ArmSetpoints to robot constants
